Fix `implicit-function-declaration` warnings, pass 6 (#656)

+++ src/melee/lb/lblanguage.h
@@ -3,11 +3,11 @@
 
 #include <dolphin/types.h>
 
-u8 lbLang_GetLanguageSetting(void);
+int lbLang_GetLanguageSetting(void);
 int lbLang_SetLanguageSetting(int language);
 int lbLang_IsSettingJP(void);
 int lbLang_IsSettingUS(void);
-u8 lbLang_GetSavedLanguage(void);
+int lbLang_GetSavedLanguage(void);
 void lbLang_SetSavedLanguage(int language);
 int lbLang_IsSavedLanguageJP(void);
 int lbLang_IsSavedLanguageUS(void);
